+++ src/main/java/gift/controller/advice/ApiGlobalExceptionHandler.java
@@ -1,7 +1,9 @@
-package gift.controller.member;
+package gift.controller.advice;
 
 import gift.exception.EmailDuplicateException;
 import gift.exception.MemberNotFoundException;
+import gift.exception.ProductNotFoundException;
+import gift.exception.WishAlreadyExistsException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
 import org.springframework.validation.FieldError;
@@ -12,13 +14,13 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@RestControllerAdvice(basePackageClasses = MemberController.class)
-public class MemberControllerAdvice {
+@RestControllerAdvice(basePackages = "gift.controller.api")
+public class ApiGlobalExceptionHandler {
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ProblemDetail handleValidationExceptions(MethodArgumentNotValidException e) {
         ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        problemDetail.setTitle("멤버 유효성 검사 실패");
+        problemDetail.setTitle("Validation fail");
 
         List<String> invalidReasonList = e.getBindingResult()
                 .getFieldErrors()
@@ -40,6 +42,20 @@ public class MemberControllerAdvice {
     @ExceptionHandler(EmailDuplicateException.class)
     public ProblemDetail handleEmailDuplicateException(EmailDuplicateException e) {
         ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
+        problemDetail.setTitle(e.getMessage());
+        return problemDetail;
+    }
+
+    @ExceptionHandler(ProductNotFoundException.class)
+    public ProblemDetail handleProductNotFoundExceptions(ProductNotFoundException e) {
+        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
+        problemDetail.setTitle(e.getMessage());
+        return problemDetail;
+    }
+
+    @ExceptionHandler(WishAlreadyExistsException.class)
+    public ProblemDetail handleWishAlreadyExistsException(WishAlreadyExistsException e) {
+        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.CONFLICT);
         problemDetail.setTitle(e.getMessage());
         return problemDetail;
     }
