--- src/main/java/gift/controller/product/ProductControllerAdvice.java
@@ -1,40 +1,0 @@
-package gift.controller.product;
-
-import gift.exception.ProductNotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@RestControllerAdvice(basePackageClasses = ProductController.class)
-public class ProductControllerAdvice {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ProblemDetail handleValidationExceptions(MethodArgumentNotValidException e) {
-        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        problemDetail.setTitle("상품 유효성 검사 실패");
-        problemDetail.setProperty("rejectedValue", e.getFieldError().getRejectedValue());
-
-        List<String> invalidReasonList = e.getBindingResult()
-                .getFieldErrors()
-                .stream()
-                .map(FieldError::getDefaultMessage)
-                .collect(Collectors.toList());
-        problemDetail.setProperty("invalidReasons", invalidReasonList);
-
-        return problemDetail;
-    }
-
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ProblemDetail handleProductNotFoundExceptions(ProductNotFoundException e) {
-        ProblemDetail problemDetail = ProblemDetail.forStatus(HttpStatus.BAD_REQUEST);
-        problemDetail.setTitle(e.getMessage());
-        return problemDetail;
-    }
-
-}
